# Introduce new trie representation